requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
black==24.8.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Django==5.1.1
django-rest-framework==0.1.0
djangorestframework==3.15.2
frozenlist==1.4.1
idna==3.10
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
propcache==0.2.0
requests==2.32.3
sqlparse==0.5.1
urllib3==2.2.3
yarl==1.14.0
